+++ demisto_sdk/commands/common/hook_validations/incident_type.py
@@ -24,10 +24,11 @@ class IncidentTypeValidator(BaseValidator):
 
         return not is_bc_broke
 
-    def is_valid_incident_type(self):
+    def is_valid_incident_type(self, validate_rn=True):
         """Check whether the Incident Type is valid or not
         """
         is_incident_type__valid = all([
+            super(IncidentTypeValidator, self).is_valid_file(validate_rn),
             self.is_valid_version()
         ])